Refactor: split people tool configs into modules
